scheduler/requirements.txt
Brotli==1.0.9
click==7.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.3
plotly==4.14.3
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
six==1.15.0
Werkzeug==1.0.1
psycopg2-binary==2.8.6
SQLAlchemy==1.4.11
requests==2.25.1
boto3==1.17.57
python-dotenv==0.17.1
